BTSTS -> MASK, added FLAGCL and FLAGST, cleaned up code
Changes mentioned are in regard to the last update via email.
MASK now uses AND instead of looping through BTST. It now preserves D1, and uses D1.B for the flag. MASK and ASCII2HEX now both take the input from D0 (and D2 for A2H) and output to D1.
Lots of comments were added to explain functionality.