UPDATES!
Added DEBUG method to set starting addr, ending addr, and file name.
Added an error check for an invalid file.
Added a manipulation method so that data is converted to the same format as it would be when added through File -> Open Data. It is missing one or two word/byte checks for something when I compare methods, but I will dissect this another time-- for now, it works fine. WHAT DOES THIS MEAN? It means we can now get HEXXED LONGS from the memory as opposed to ASCII data. Get rid of those ASCII2HEX conversions!