Thanksgiving Update
MOVE completely implemented except for the OP modes we never use.
Most code can be translated to the screen via mode and register.
Added a new compare method CMPMSK which compares both 0 bits and 1 bits. 1 in D2 when bit should be 1, 1 in D3 when bit should be 0. Only currently implemented for MOVEQ which is outside of our instruction-- but useful to see how the compiler is translating our MOVE code.
IO will display comma when necessary.
BAD will display best guess for OP-code if the OP-code was recognized before it went bad.
Added and changed way too many things to list.