1:1 Instructions
Instructions now use the new compare method, so that all 0s, 1s, and don't cares are included. This means that 0 and 1 constants no longer need to be checked for BAD within the instruction branches.

Also added some extra instructions that we will not use, but will be recognized and displayed as BAD but with instruction name.